pyproject.toml
[tool.poetry]
name = "backbone"
version = "0.3.0"
description = "Tools for development with Backbone"
authors = ["Backbone Developers <inbox@backbone.dev>"]
license = "Proprietary"
packages = [{ include = "backbone" }]

[tool.poetry.scripts]
backbone = 'backbone:main'

[tool.poetry.dependencies]
python = "^3.7"
PyNaCl = "^1.4.0"
typer = "^0.3.2"
pydantic = "^1.8.2"
halo = "^0.0.31"
httpx = "^0.18.2"
getmac = "^0.8.2"
cbor2 = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
pytest-asyncio = "^0.15.1"
black = "^21.7b0"
isort = "^5.9.3"
autoflake = "^1.4"

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
